refactor(web): create separate accordion functions for has and fpa
