+++ appeals/web/src/server/appeals/appeal-details/accordions/index.js
@@ -1,0 +1,2 @@
+export { generateAccordion as generateHasAccordion } from './has.js';
+export { generateAccordion as generateFpaAccordion } from './fpa.js';
