+++ appeals/web/src/server/appeals/appeal-details/accordions/utils/remove-accordion-components-actions.js
@@ -1,0 +1,30 @@
+import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+
+const caseDocumentationTableActionColumnIndex = 3;
+
+/**
+ * @param {PageComponent[]} accordionComponents
+ * @returns {void}
+ */
+export function removeAccordionComponentsActions(accordionComponents) {
+	accordionComponents.forEach((component) => {
+		switch (component.type) {
+			case 'summary-list':
+				component.parameters.rows = component.parameters.rows.map(
+					(/** @type {SummaryListRowProperties} */ row) => removeSummaryListActions(row)
+				);
+				break;
+			case 'table':
+				component.parameters.rows.forEach((/** @type {TableCellProperties[]} */ row) =>
+					row.forEach((cell, index) => {
+						if (index === caseDocumentationTableActionColumnIndex && 'html' in cell) {
+							cell.html = '';
+						}
+					})
+				);
+				break;
+			default:
+				break;
+		}
+	});
+}
