+++ appeals/web/src/server/appeals/appeal-details/appeal-details.mapper.js
@@ -1,28 +1,21 @@
-import logger from '../../lib/logger.js';
-import {
-	initialiseAndMapAppealData,
-	generateDecisionDocumentDownloadHtml
-} from '#lib/mappers/appeal/appeal.mapper.js';
-import { buildNotificationBanners } from '#lib/mappers/notification-banners.mapper.js';
-import { isDefined } from '#lib/ts-utilities.js';
-import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
-import { preRenderPageComponents } from '#lib/nunjucks-template-builders/page-component-rendering.js';
-import { appealShortReference } from '#lib/appeals-formatter.js';
-import { mapVirusCheckStatus } from '#appeals/appeal-documents/appeal-documents.mapper.js';
-import { getAppealTypesFromId } from '#appeals/appeal-details/change-appeal-type/change-appeal-type.service.js';
-import { APPEAL_VIRUS_CHECK_STATUS } from 'pins-data-model';
-import { APPEAL_CASE_STATUS } from 'pins-data-model';
-import { APPEAL_TYPE, FEATURE_FLAG_NAMES } from '@pins/appeals/constants/common.js';
-import { isFeatureActive } from '#common/feature-flags.js';
-import { addNotificationBannerToSession } from '#lib/session-utilities.js';
-import {
-	generateIssueDecisionUrl,
-	generateStartTimetableUrl
-} from './issue-decision/issue-decision.mapper.js';
-import { dateISOStringToDisplayDate, getTodaysISOString } from '#lib/dates.js';
 import { caseNotesWithMappedUsers } from '#appeals/appeal-details/case-notes/case-notes.formatter.js';
-import { userHasPermission } from '#lib/mappers/permissions.mapper.js';
-import { permissionNames } from '#environment/permissions.js';
+import { getAppealTypesFromId } from '#appeals/appeal-details/change-appeal-type/change-appeal-type.service.js';
+import { mapVirusCheckStatus } from '#appeals/appeal-documents/appeal-documents.mapper.js';
+import { isFeatureActive } from '#common/feature-flags.js';
+import { appealShortReference } from '#lib/appeals-formatter.js';
+import { dateISOStringToDisplayDate, getTodaysISOString } from '#lib/dates.js';
+import {
+	generateDecisionDocumentDownloadHtml,
+	initialiseAndMapAppealData
+} from '#lib/mappers/appeal/appeal.mapper.js';
+import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { buildNotificationBanners } from '#lib/mappers/notification-banners.mapper.js';
+import { preRenderPageComponents } from '#lib/nunjucks-template-builders/page-component-rendering.js';
+import { isDefined } from '#lib/ts-utilities.js';
+import { APPEAL_TYPE, FEATURE_FLAG_NAMES } from '@pins/appeals/constants/common.js';
+import { APPEAL_CASE_STATUS, APPEAL_VIRUS_CHECK_STATUS } from 'pins-data-model';
+import logger from '../../lib/logger.js';
+import { generateFpaAccordion, generateHasAccordion } from './accordions/index.js';
 
 export const pageHeading = 'Case details';
 
@@ -252,127 +245,6 @@ export async function appealDetailsPage(
 }
 
 /**
- * @param {import('./appeal-details.types.js').WebAppeal} appealDetails
- * @param {import('express-session').Session & Partial<import('express-session').SessionData>} session
- * @param {PageComponent[]} accordionComponents
- * @param {boolean} [ipCommentsAwaitingReview]
- * @returns {void}
- */
-function mapStatusDependentNotifications(
-	appealDetails,
-	session,
-	accordionComponents,
-	ipCommentsAwaitingReview = false
-) {
-	switch (appealDetails.appealStatus) {
-		case APPEAL_CASE_STATUS.ASSIGN_CASE_OFFICER:
-			addNotificationBannerToSession(
-				session,
-				'assignCaseOfficer',
-				appealDetails.appealId,
-				`<p class="govuk-notification-banner__heading">Appeal ready to be assigned to case officer</p><p><a class="govuk-notification-banner__link" href="/appeals-service/appeal-details/${appealDetails.appealId}/assign-user/case-officer" data-cy="banner-assign-case-officer">Assign case officer</a></p>`
-			);
-			break;
-		case APPEAL_CASE_STATUS.VALIDATION:
-			addNotificationBannerToSession(
-				session,
-				'readyForValidation',
-				appealDetails.appealId,
-				`<p class="govuk-notification-banner__heading">Appeal ready for validation</p><p><a class="govuk-notification-banner__link" data-cy="validate-appeal" href="/appeals-service/appeal-details/${appealDetails.appealId}/appellant-case">Validate <span class="govuk-visually-hidden">appeal</span></a></p>`
-			);
-			break;
-		case APPEAL_CASE_STATUS.LPA_QUESTIONNAIRE:
-			if (appealDetails.lpaQuestionnaireId) {
-				addNotificationBannerToSession(
-					session,
-					'readyForLpaQuestionnaireReview',
-					appealDetails.appealId,
-					`<p class="govuk-notification-banner__heading">LPA questionnaire ready for review</p><p><a class="govuk-notification-banner__link" data-cy="review-lpa-questionnaire-banner" href="/appeals-service/appeal-details/${appealDetails.appealId}/lpa-questionnaire/${appealDetails.lpaQuestionnaireId}">Review <span class="govuk-visually-hidden">LPA questionnaire</span></a></p>`
-				);
-			}
-			break;
-		case APPEAL_CASE_STATUS.ISSUE_DETERMINATION:
-			addNotificationBannerToSession(
-				session,
-				'readyForDecision',
-				appealDetails.appealId,
-				`<p class="govuk-notification-banner__heading">Ready for decision</p><p><a class="govuk-notification-banner__link" data-cy="issue-determination" href="${generateIssueDecisionUrl(
-					appealDetails.appealId
-				)}">Issue decision</a></p>`
-			);
-			break;
-		case APPEAL_CASE_STATUS.READY_TO_START:
-			addNotificationBannerToSession(
-				session,
-				'appealValidAndReadyToStart',
-				appealDetails.appealId,
-				`<p class="govuk-notification-banner__heading">Appeal valid</p><p><a class="govuk-notification-banner__link" data-cy="ready-to-start" href="${generateStartTimetableUrl(
-					appealDetails.appealId
-				)}">Start case</a></p>`
-			);
-			break;
-		case APPEAL_CASE_STATUS.AWAITING_TRANSFER:
-			addNotificationBannerToSession(
-				session,
-				'appealAwaitingTransfer',
-				appealDetails.appealId,
-				`<p class="govuk-notification-banner__heading">This appeal is awaiting transfer</p><p class="govuk-body">The appeal must be transferred to Horizon. When this is done, <a class="govuk-link" data-cy="awaiting-transfer" href="/appeals-service/appeal-details/${appealDetails.appealId}/change-appeal-type/add-horizon-reference">update the appeal with the new horizon reference</a>.</p>`
-			);
-			removeAccordionComponentsActions(accordionComponents);
-			break;
-		case APPEAL_CASE_STATUS.STATEMENTS:
-			if (ipCommentsAwaitingReview) {
-				addNotificationBannerToSession(
-					session,
-					'interestedPartyCommentsAwaitingReview',
-					appealDetails.appealId,
-					`<p class="govuk-notification-banner__heading">Review comments</p><p><a class="govuk-notification-banner__link" href="/appeals-service/appeal-details/${appealDetails.appealId}/interested-party-comments" data-cy="banner-review-ip-comments">Review <span class="govuk-visually-hidden">interested party comments</span></a></p>`
-				);
-			}
-			break;
-		default:
-			break;
-	}
-
-	if (
-		'notificationBanners' in session &&
-		'appealAwaitingTransfer' in session.notificationBanners &&
-		appealDetails.appealStatus !== 'awaiting_transfer'
-	) {
-		delete session.notificationBanners.appealAwaitingTransfer;
-	}
-}
-
-const caseDocumentationTableActionColumnIndex = 3;
-
-/**
- * @param {PageComponent[]} accordionComponents
- * @returns {void}
- */
-function removeAccordionComponentsActions(accordionComponents) {
-	accordionComponents.forEach((component) => {
-		switch (component.type) {
-			case 'summary-list':
-				component.parameters.rows = component.parameters.rows.map(
-					(/** @type {SummaryListRowProperties} */ row) => removeSummaryListActions(row)
-				);
-				break;
-			case 'table':
-				component.parameters.rows.forEach((/** @type {TableCellProperties[]} */ row) =>
-					row.forEach((cell, index) => {
-						if (index === caseDocumentationTableActionColumnIndex && 'html' in cell) {
-							cell.html = '';
-						}
-					})
-				);
-				break;
-			default:
-				break;
-		}
-	});
-}
-
-/**
  *
  * @param {import('./appeal-details.types.js').WebAppeal} appealDetails
  * @param {{appeal: MappedInstructions}} mappedData
@@ -383,253 +255,13 @@ function removeAccordionComponentsActions(accordionComponents) {
 function generateAccordionItems(appealDetails, mappedData, session, ipCommentsAwaitingReview) {
 	switch (appealDetails.appealType) {
 		case APPEAL_TYPE.D:
-			return generateAccordion(appealDetails, mappedData, session);
+			return generateHasAccordion(appealDetails, mappedData, session);
 		case APPEAL_TYPE.W:
 			if (!isFeatureActive(FEATURE_FLAG_NAMES.SECTION_78)) {
 				throw new Error('Feature flag inactive for S78');
 			}
-			return generateAccordion(appealDetails, mappedData, session, ipCommentsAwaitingReview);
+			return generateFpaAccordion(appealDetails, mappedData, session, ipCommentsAwaitingReview);
 		default:
 			throw new Error('Invalid appealType, unable to generate display page');
 	}
-}
-
-/**
- *
- * @param {import('./appeal-details.types.js').WebAppeal} appealDetails
- * @param {{appeal: MappedInstructions}} mappedData
- * @param {import('express-session').Session & Partial<import('express-session').SessionData>} session
- * @param {boolean} [ipCommentsAwaitingReview]
- * @returns
- */
-function generateAccordion(appealDetails, mappedData, session, ipCommentsAwaitingReview) {
-	/** @type {PageComponent} */
-	const caseOverview = {
-		type: 'summary-list',
-		parameters: {
-			rows: [
-				mappedData.appeal.appealType.display.summaryListItem,
-				removeSummaryListActions(mappedData.appeal?.caseProcedure?.display.summaryListItem),
-				mappedData.appeal?.linkedAppeals?.display.summaryListItem,
-				mappedData.appeal?.otherAppeals?.display.summaryListItem,
-				mappedData.appeal?.allocationDetails?.display.summaryListItem,
-				removeSummaryListActions(mappedData.appeal?.lpaReference?.display.summaryListItem),
-				mappedData.appeal?.decision?.display.summaryListItem
-			].filter(isDefined)
-		}
-	};
-
-	/** @type {PageComponent} */
-	const siteDetails = {
-		type: 'summary-list',
-		parameters: {
-			rows: [
-				mappedData.appeal.lpaInspectorAccess.display.summaryListItem,
-				mappedData.appeal.appellantInspectorAccess.display.summaryListItem,
-				mappedData.appeal.neighbouringSiteIsAffected.display.summaryListItem,
-				mappedData.appeal.lpaNeighbouringSites.display.summaryListItem,
-				mappedData.appeal.inspectorNeighbouringSites.display.summaryListItem,
-				mappedData.appeal.lpaHealthAndSafety.display.summaryListItem,
-				mappedData.appeal.appellantHealthAndSafety.display.summaryListItem,
-				mappedData.appeal.visitType.display.summaryListItem
-			].filter(isDefined)
-		}
-	};
-
-	/** @type {PageComponent[]} */
-	const caseTimetable = (() => {
-		switch (appealDetails.appealType) {
-			case APPEAL_TYPE.D:
-				return appealDetails.startedAt
-					? [
-							{
-								type: 'summary-list',
-								parameters: {
-									rows: [
-										mappedData.appeal.validAt.display.summaryListItem,
-										mappedData.appeal.startedAt.display.summaryListItem,
-										mappedData.appeal.lpaQuestionnaireDueDate.display.summaryListItem,
-										mappedData.appeal.siteVisitDate.display.summaryListItem
-									].filter(isDefined)
-								}
-							}
-					  ]
-					: [
-							{
-								type: 'summary-list',
-								parameters: {
-									rows: [
-										mappedData.appeal.validAt.display.summaryListItem,
-										mappedData.appeal.startedAt.display.summaryListItem
-									].filter(isDefined)
-								}
-							}
-					  ];
-			case APPEAL_TYPE.W:
-				if (!isFeatureActive(FEATURE_FLAG_NAMES.SECTION_78)) {
-					throw new Error('Feature flag inactive for S78');
-				}
-				return [
-					{
-						type: 'summary-list',
-						parameters: {
-							rows: [
-								mappedData.appeal.validAt.display.summaryListItem,
-								mappedData.appeal.startedAt.display.summaryListItem,
-								mappedData.appeal.lpaQuestionnaireDueDate.display.summaryListItem,
-								mappedData.appeal.lpaStatementDueDate.display.summaryListItem,
-								mappedData.appeal.ipCommentsDueDate.display.summaryListItem,
-								mappedData.appeal.appellantFinalCommentDueDate.display.summaryListItem,
-								mappedData.appeal.lpaFinalCommentDueDate.display.summaryListItem,
-								mappedData.appeal.s106ObligationDueDate.display.summaryListItem
-							].filter(isDefined)
-						}
-					}
-				];
-			default:
-				throw new Error('Invalid appealType, unable to generate display page');
-		}
-	})();
-
-	/** @type {PageComponent} */
-	const caseDocumentation = {
-		type: 'table',
-		parameters: {
-			head: [
-				{ text: 'Documentation' },
-				{ text: 'Status' },
-				{ text: 'Received' },
-				{ text: 'Action' }
-			],
-			rows: [
-				mappedData.appeal.appellantCase.display.tableItem,
-				mappedData.appeal.lpaQuestionnaire.display.tableItem,
-				mappedData.appeal.appealDecision.display.tableItem
-			].filter(isDefined),
-			firstCellIsHeader: true
-		}
-	};
-
-	/** @type {PageComponent} */
-	const caseCosts = {
-		type: 'table',
-		parameters: {
-			head: [{ text: 'Documentation' }, { text: 'Status' }, { text: 'Action' }],
-			rows: [
-				mappedData.appeal.costsAppellantApplication.display.tableItem,
-				mappedData.appeal.costsAppellantWithdrawal.display.tableItem,
-				mappedData.appeal.costsAppellantCorrespondence.display.tableItem,
-				mappedData.appeal.costsLpaApplication.display.tableItem,
-				mappedData.appeal.costsLpaWithdrawal.display.tableItem,
-				mappedData.appeal.costsLpaCorrespondence.display.tableItem,
-				mappedData.appeal.costsDecision.display.tableItem
-			].filter(isDefined),
-			firstCellIsHeader: true
-		}
-	};
-
-	/** @type {PageComponent} */
-	const caseContacts = {
-		type: 'summary-list',
-		parameters: {
-			rows: [
-				mappedData.appeal.appellant.display.summaryListItem,
-				mappedData.appeal.agent.display.summaryListItem,
-				removeSummaryListActions({
-					...mappedData.appeal.localPlanningAuthority.display.summaryListItem,
-					key: {
-						text: 'LPA'
-					}
-				})
-			].filter(isDefined)
-		}
-	};
-
-	/** @type {PageComponent} */
-	const caseTeam = {
-		type: 'summary-list',
-		parameters: {
-			rows: [
-				mappedData.appeal.caseOfficer.display.summaryListItem,
-				mappedData.appeal.inspector.display.summaryListItem
-			].filter(isDefined)
-		}
-	};
-
-	/** @type {PageComponent} */
-	const caseManagement = {
-		type: 'summary-list',
-		parameters: {
-			rows: [
-				mappedData.appeal.crossTeamCorrespondence.display.summaryListItem,
-				mappedData.appeal.inspectorCorrespondence.display.summaryListItem,
-				mappedData.appeal.caseHistory.display.summaryListItem,
-				mappedData.appeal.appealWithdrawal.display.summaryListItem
-			]
-		}
-	};
-
-	const accordionComponents = [
-		caseOverview,
-		siteDetails,
-		caseTimetable[0],
-		caseDocumentation,
-		caseContacts,
-		caseTeam
-	];
-
-	mapStatusDependentNotifications(
-		appealDetails,
-		session,
-		accordionComponents,
-		ipCommentsAwaitingReview
-	);
-
-	if (!userHasPermission(permissionNames.viewCaseDetails, session)) {
-		removeAccordionComponentsActions(accordionComponents);
-	}
-
-	/** @type {PageComponent} */
-	const appealDetailsAccordion = {
-		type: 'accordion',
-		parameters: {
-			id: 'accordion-default' + appealDetails.appealId,
-			items: [
-				{
-					heading: { text: 'Overview' },
-					content: { html: '', pageComponents: [caseOverview] }
-				},
-				{
-					heading: { text: 'Site' },
-					content: { html: '', pageComponents: [siteDetails] }
-				},
-				{
-					heading: { text: 'Timetable' },
-					content: { html: '', pageComponents: caseTimetable }
-				},
-				{
-					heading: { text: 'Documentation' },
-					content: { html: '', pageComponents: [caseDocumentation] }
-				},
-				{
-					heading: { text: 'Costs' },
-					content: { html: '', pageComponents: [caseCosts] }
-				},
-				{
-					heading: { text: 'Contacts' },
-					content: { html: '', pageComponents: [caseContacts] }
-				},
-				{
-					heading: { text: 'Team' },
-					content: { html: '', pageComponents: [caseTeam] }
-				},
-				{
-					heading: { text: 'Case management' },
-					content: { html: '', pageComponents: [caseManagement] }
-				}
-			]
-		}
-	};
-
-	return appealDetailsAccordion;
 }
